+++ RELEASE-NOTES.md
@@ -10,6 +10,7 @@ Under development
 * Dropped support for PHP older than 7.1
 * Dropped support for MediaWiki older than 1.31
 * Added API key support for Leaflet layers via the `egMapsLeafletLayersApiKeys` setting (by Gilles Cébélieu)
+* Updated Leaflet library from version 1.3.1 to version 1.3.4
 
 ### GeoJson support